texture atlas layout in the texture atlas module
